add Intro component
