+++ lib/components/MyCrossword/MyCrossword.tsx
@@ -3,10 +3,9 @@ import { CellChange, CellFocus, GuardianCrossword, GuessGrid } from '~/types';
 import { DEFAULT_CELL_MATCHER, DEFAULT_HTML_TAGS } from '~/utils/general';
 import classNames from 'classnames';
 import Crossword from '~/components/Crossword/Crossword';
-import React, { useEffect, useRef, useState } from 'react';
+import React, { useState } from 'react';
+import Intro from '~/components/Intro/Intro';
 import './MyCrossword.css';
-
-type Timeout = ReturnType<typeof setInterval>;
 
 type Theme =
   | 'red'
@@ -67,76 +66,10 @@ export default function MyCrossword({
 
   // Intro state
   const [showIntro, setShowIntro] = useState(intro !== undefined);
-  const [countdown, setCountdown] = useState(
-    intro?.timeout !== undefined ? Math.ceil(intro.timeout / 1000) : 0,
-  );
-  const [canContinue, setCanContinue] = useState(intro?.timeout === undefined);
-  const timerRef = useRef<Timeout | null>(null);
 
-  const handleClearTimer = () => {
-    if (timerRef.current !== null) {
-      clearInterval(timerRef.current);
-      timerRef.current = null;
-    }
-  };
-
-  useEffect(() => {
-    if (intro === undefined || intro.timeout === undefined) {
-      setCanContinue(true);
-      return;
-    }
-
-    setCountdown(Math.ceil(intro.timeout / 1000));
-    setShowIntro(true);
-    setCanContinue(false);
-
-    handleClearTimer();
-    const start = Date.now();
-    timerRef.current = setInterval(() => {
-      const elapsed = Date.now() - start;
-      const left = Math.max(0, intro.timeout! - elapsed);
-
-      setCountdown(Math.ceil(left / 1000));
-
-      if (left <= 0) {
-        setCanContinue(true);
-        handleClearTimer();
-      }
-    }, 250);
-
-    return () => {
-      handleClearTimer();
-    };
-  }, [intro?.timeout, intro?.node, id]);
-
-  const handleSkip = () => {
-    if (!canContinue) return;
+  const handleIntroContinue = () => {
     setShowIntro(false);
-    handleClearTimer();
   };
-
-  function CircleCounter({ count, ready }: { count: number; ready: boolean }) {
-    return (
-      <div
-        style={{
-          width: 36,
-          height: 36,
-          borderRadius: '50%',
-          border: '2px solid #1976d2',
-          display: 'flex',
-          alignItems: 'center',
-          justifyContent: 'center',
-          fontWeight: 600,
-          fontSize: 15,
-          color: '#1976d2',
-          background: '#fff',
-          flexShrink: 0,
-        }}
-      >
-        {ready ? '›' : count}
-      </div>
-    );
-  }
 
   if (showIntro && intro) {
     return (
@@ -146,39 +79,12 @@ export default function MyCrossword({
           className,
           'MyCrossword-intro',
         )}
-        style={{
-          display: 'flex',
-          flexDirection: 'column',
-          alignItems: 'center',
-          justifyContent: 'center',
-          minHeight: 320,
-          padding: 24,
-        }}
       >
-        <div style={{ marginBottom: 32 }}>{intro.node}</div>
-        <button
-          type="button"
-          onClick={handleSkip}
-          disabled={!canContinue}
-          style={{
-            display: 'flex',
-            alignItems: 'center',
-            gap: 10,
-            background: 'none',
-            border: 'none',
-            cursor: canContinue ? 'pointer' : 'default',
-            padding: 0,
-            fontSize: 16,
-            color: 'inherit',
-            opacity: canContinue ? 1 : 0.6,
-          }}
-          aria-label={
-            canContinue ? 'Continue to crossword' : `Skip in ${countdown}`
-          }
-        >
-          <span>{canContinue ? 'Continue to crossword' : 'Skip in'}</span>
-          <CircleCounter count={countdown} ready={canContinue} />
-        </button>
+        <Intro
+          node={intro.node}
+          timeout={intro.timeout}
+          onContinue={handleIntroContinue}
+        />
       </div>
     );
   }
